fix bug with casting to wrong type

+++ ConceptEngine/ConceptEngine/ConceptEngineFramework/Graphics/DirectX12/Resources/CECubeRenderTarget.cpp
@@ -12,8 +12,8 @@ CECubeRenderTarget::CECubeRenderTarget(ID3D12Device* device,
 	m_height = height;
 	Format = format;
 
-	m_viewport = {0.0f, 0.0f, (float)width, (float)height, 0.0f, 1.0f};
-	m_scissorRect = {0, 0, (float)width, (float)height};
+	m_viewport = {0.0f, 0.0f, static_cast<float>(width), static_cast<float>(height), 0.0f, 1.0f};
+	m_scissorRect = {0, 0, (long)width, (long)height};
 }
 
 ID3D12Resource* CECubeRenderTarget::Resource() {
